+++ tests/unit/nodes/physics_engine/test_example.py
@@ -1,0 +1,3 @@
+def test_function():
+    assert(True)
+
